SPT - scorelevel 기능 함수 분할
기존에는 scorelevel 관련 기능 함수가 테트리스_완성-점수처리.c 파일에 선언되어 있어서 코드가 길어져 가독성이 떨어짐.
따라서 scorelevel.c 파일과 scorelevel.h 파일을 생성한 후 scorelevel 관련 기능 함수를 선언함.
