+++ scorelevel.h
@@ -1,0 +1,12 @@
+#ifndef __SCORELEVEL_H__
+
+#define __SCORELEVEL_H__
+
+int static score = 0;
+int static level = 1;
+int static speed = 180;
+
+void Print_scorelevel(void);
+void Count_score(void);
+
+#endif
